fix: Support `riff-raff.yaml` files without top-level `stacks`
A valid `riff-raff.yaml` file does not need to have `stacks` defined at the top level;
it can be defined per deployment instead.

We currently use the value of stacks to create a default project name.

In this change we require `stacks` to have exactly 1 item for this, else require the `projectName` input to be set.