feat!: Remove the `app` input
BREAKING CHANGE: Remove `app` input in favour of `projectName`

The `app` input is used to build half of the Riff-Raff project name.
The other half comes from the `stacks` listed in the `riff-raff.yaml` file.

This is quite complex, especially when compared to the `projectName` input
which is the Riff-Raff project name verbatim.

To migrate, replace the `app` input with the Riff-Raff project name.
If this is not known, the build logs can be used.
https://github.com/guardian/riffraff-platform can be used also.