chore(deps): Revert `@aws-sdk/client-s3` back to v3.703.0
Since the upgrade, we've experienced `RequestTimeout` errors in the build of a `guardian/dotcom-rendering`.
We patched this in https://github.com/guardian/actions-riff-raff/releases/tag/v4.1.4.

Today, we started seeing `guardian/ophan` experience the same issues.
This means the randomly chosen 10 second timeout applied in https://github.com/guardian/actions-riff-raff/pull/191
doesn't work for all cases.

This change "reverts" the `@aws-sdk/client-s3` and `@aws-sdk/credential-providers` libraries back to known working versions.

Note: it isn't a revert in the git sense as the changes are interleaved with others.
Additionally, we're now pinning the version number, whereas previously it was ranged.